refactor: redirect 관련 코드 web 모듈로 이동
