--- module-domain/src/main/java/inspiration/utils/SecurityUtil.java
@@ -1,21 +1,0 @@
-package inspiration.utils;
-
-import inspiration.exception.UnauthorizedAccessRequestException;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-
-@Slf4j
-public class SecurityUtil {
-    private SecurityUtil() { }
-
-    public static Long getCurrentMemberId() {
-        final Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-
-        if(authentication == null || authentication.getName() == null) {
-            throw new UnauthorizedAccessRequestException("Security Context에 인증 정보가 없습니다.");
-        }
-
-        return Long.parseLong(authentication.getName());
-    }
-}
