+++ module-web/src/main/java/inspiration/redirect/PasswordAuthRedirectViewUtil.java
@@ -1,4 +1,4 @@
-package inspiration.utils;
+package inspiration.redirect;
 
 import org.springframework.web.servlet.view.RedirectView;
